move themeLegendSwitch
