+++ lib/ui/elements/_elements.mjs
@@ -28,7 +28,6 @@ The module exports a collection of ui element modules.
 @requires module:/ui/elements/searchbox
 @requires module:/ui/elements/slider
 @requires module:/ui/elements/slider_ab
-@requires module:/ui/elements/themeLegendSwitch
 @requires module:/ui/elements/toast
 @requires module:/ui/elements/tooltip
 @requires module:/ui/elements/userLocale
@@ -59,7 +58,6 @@ import reportLink from './reportLink.mjs';
 import searchbox from './searchbox.mjs';
 import slider from './slider.mjs';
 import slider_ab from './slider_ab.mjs';
-import themeLegendSwitch from './themeLegendSwitch.mjs';
 import toast from './toast.mjs';
 import tooltip from './tooltip.mjs';
 import userLocale from './userLocale.mjs';
@@ -91,7 +89,6 @@ export default {
   searchbox,
   slider,
   slider_ab,
-  themeLegendSwitch,
   toast,
   tooltip,
   userLocale,
