+++ lib/ui/layers/legends/categorized.mjs
@@ -4,12 +4,16 @@
 The categorized theme legend module exports the categorizedTheme to the `ui.layers.legend{}` library object.
 
 @requires /ui/layers/legends/utils
-@requires /ui/elements/themeLegendSwitch
 
 @module /ui/layers/legends/categorized
 */
 
-import { catElement, clusterLegend, themeParser } from './utils.mjs';
+import {
+  catElement,
+  clusterLegend,
+  themeLegendSwitch,
+  themeParser,
+} from './utils.mjs';
 
 /**
 @function categorizedTheme
@@ -27,7 +31,7 @@ export default function categorizedTheme(layer) {
   const theme = layer.style.theme;
 
   // Switch all control
-  theme.legend.switch = layer.filter && mapp.ui.elements.themeLegendSwitch();
+  theme.legend.switch = layer.filter && themeLegendSwitch();
 
   theme.categories.forEach((cat) => {
     catElement(cat, theme, layer);
