+++ lib/ui/layers/legends/distributed.mjs
@@ -54,6 +54,7 @@ export default function distributedTheme(layer) {
     layer.style.legend.replaceChildren(...theme.legend.node.children);
   }
 
+  // The method to rebuild the legend must be called after each render is complete.
   layer.L.once('postrender', () => {
     if (theme.key !== layer.style.theme.key) return;
