+++ lib/ui/layers/legends/graduated.mjs
@@ -4,10 +4,12 @@
 The graduated theme legend module exports the graduatedTheme to the `ui.layers.legend{}` library object.
 
 @requires /ui/elements/legendIcon
-@requires /ui/elements/themeLegendSwitch
+@requires /ui/layers/legends/utils
 
 @module /ui/layers/legends/graduated
 */
+
+import { themeLegendSwitch } from './utils.mjs';
 
 /**
 @function graduatedTheme
@@ -25,7 +27,7 @@ export default function graduatedTheme(layer) {
   theme.legend ??= {};
 
   // Switch all control
-  const switchAll = mapp.ui.elements.themeLegendSwitch();
+  const switchAll = themeLegendSwitch();
 
   const catElements = theme.categories
     .filter((cat) => cat.value !== undefined)
