--- tests/lib/ui/elements/themeLegendSwitch.test.mjs
@@ -1,2 +1,0 @@
-//TODO in v4.13.0
-export function themeLegendSwitch() {}
